More explicit dependencies to help with test support library.
