+++ Sources/SQLiteData/CloudKit/CloudKit+StructuredQueries.swift
@@ -1,7 +1,6 @@
 #if canImport(CloudKit)
   import CloudKit
   import CryptoKit
-  import CustomDump
   import StructuredQueriesCore
 
   extension _CKRecord where Self == CKRecord {
