+++ Sources/SQLiteData/CloudKit/Internal/RecordType.swift
@@ -1,4 +1,3 @@
-import CustomDump
 
 @Table("sqlitedata_icloud_recordTypes")
 package struct RecordType: Hashable {
@@ -7,18 +6,4 @@ package struct RecordType: Hashable {
   package let schema: String
   @Column(as: Set<TableInfo>.JSONRepresentation.self)
   package let tableInfo: Set<TableInfo>
-}
-
-extension RecordType: CustomDumpReflectable {
-  package var customDumpMirror: Mirror {
-    Mirror(
-      self,
-      children: [
-        ("tableName", tableName as Any),
-        ("schema", schema),
-        ("tableInfo", tableInfo.sorted(by: { $0.name < $1.name })),
-      ],
-      displayStyle: .struct
-    )
-  }
 }
